moved map controllers to a new folder 'src/Infrastructure/Controller/Map/'
