+++ src/Infrastructure/Controller/Map/MapController.php
@@ -2,11 +2,10 @@
 
 declare(strict_types=1);
 
-namespace App\Infrastructure\Controller;
+namespace App\Infrastructure\Controller\Map;
 
 use App\Domain\Regulation\Repository\LocationRepositoryInterface;
 use App\Infrastructure\Form\Map\MapFilterFormType;
-use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\Form\FormFactoryInterface;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -14,7 +13,7 @@ use Symfony\Component\HttpKernel\Attribute\MapQueryParameter;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\Routing\RouterInterface;
 
-final class GetMapFilterController extends AbstractController
+final class MapController
 {
     public function __construct(
         private \Twig\Environment $twig,
@@ -25,8 +24,8 @@ final class GetMapFilterController extends AbstractController
     }
 
     #[Route(
-        '/_get_map_filter',
-        name: 'get_map_filter',
+        '/carte',
+        name: 'app_carto',
         methods: ['GET'],
     )]
     public function __invoke(Request $request, #[MapQueryParameter] array $map_filter_form = ['category' => 'permanents_and_temporaries', 'display_future_regulations' => 'no', 'display_past_regulations' => 'no']): Response
@@ -41,7 +40,6 @@ final class GetMapFilterController extends AbstractController
                 ],
             ],
         );
-
         $form->handleRequest($request); // auto-fill the form with the query parameters from the URL
 
         // the array '$map_filter_form' can be defined without the 'category' key for example, so we have to set a default value eventually
@@ -53,7 +51,7 @@ final class GetMapFilterController extends AbstractController
 
         return new Response(
             $this->twig->render(
-                name: 'map/map_filter.html.twig',
+                name: 'map/map.html.twig',
                 context: [
                     'locationsAsGeoJson' => $locationsAsGeoJson,
                     'form' => $form->createView(),
